Tue, Dec 20, 2022, 2:12 PM +00:00

+++ readability/readability.c
@@ -24,8 +24,8 @@ int main(void)
     int sentences = count_sentences(input);
     printf("sentences : %i\n", sentences);
 
-    int index = (0.0588 * ((letters / words) * 100)) - (0.296 * ((sentences / words) * 100)) - 15.8;
-    printf("Grade %i\n", index);
+    int index = (0.0588 * (((float) letters / (float) words) * 100)) - (0.296 * (((float) sentences / (float) words) * 100)) - 15.8;
+    printf("Grade %d\n", index);
 
 
 }
